requirements.txt
argon2-cffi==19.2.0
Authlib==0.13
cffi==1.14.0
click==7.1.1
cryptography==2.8
Flask==1.1.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.20
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
WTForms==2.2.1
